+++ back-end/router/community/liked_history.js
@@ -1,0 +1,19 @@
+const likedHistoryRouter = require('express').Router()
+const UserInfo = require('../../model/userInfo/userInfo')
+
+likedHistoryRouter.get('/', async (req, res) => {
+    let ret = {}
+
+    // retrieve user info from database
+    await UserInfo.findOne({ user_name: req.user.username }, (err, result) => {
+        if (err) {
+            console.error(err)
+        } else {
+            // extract being liked history
+            ret = result.others_liked_history
+            res.json(ret)
+            console.log(ret)
+        }
+    })
+})
+module.exports = likedHistoryRouter
